refactor: Rename files tftp_common* to tftpCommon* (u-t OK!)
Github issue #4
