+++ source/tests/tftpSrv_test.cpp
@@ -5,7 +5,7 @@
 #include <type_traits>
 
 
-#include "../tftp_common.h"
+#include "../tftpCommon.h"
 #include "../tftpSrv.h"
 #include "test.h"
