+++ source/tftpBase.h
@@ -17,7 +17,7 @@
 
 #include <shared_mutex>
 
-#include "tftp_common.h"
+#include "tftpCommon.h"
 #include "tftpSettings.h"
 
 namespace tftp
